crawler/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
deep-translator==1.11.4
# feedparser removed due to cgi module deprecation in python 3.13+
